requirements.txt
asgiref==3.2.10
Django==3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-taggit==1.3.0
easy-thumbnails==2.7
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
sqlparse==0.3.1
